refactor: 카프가 관련 패키지 리팩토링
