+++ src/main/java/ChatManagement/chat/application/ChatMessageService.java
@@ -7,10 +7,10 @@ import ChatManagement.chat.repository.ChatRoomRepository;
 import ChatManagement.chat.presentation.dto.ChatRoomRequest;
 import ChatManagement.chat.presentation.dto.ChatMessageResponse;
 import ChatManagement.global.execption.NotFoundChatRoomException;
-import ChatManagement.kafka.service.dto.KafkaMessage;
-import ChatManagement.kafka.service.dto.LogMessage;
-import ChatManagement.kafka.service.KafkaService;
-import ChatManagement.kafka.service.dto.type.LogType;
+import ChatManagement.kafka.application.dto.KafkaMessage;
+import ChatManagement.kafka.application.dto.LogMessage;
+import ChatManagement.kafka.application.KafkaService;
+import ChatManagement.kafka.application.dto.type.LogType;
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
