+++ src/main/java/ChatManagement/global/config/KafkaConfig.java
@@ -1,6 +1,6 @@
-package ChatManagement.kafka.config;
+package ChatManagement.global.config;
 
-import ChatManagement.kafka.service.dto.KafkaMessage;
+import ChatManagement.kafka.application.dto.KafkaMessage;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.extern.slf4j.Slf4j;
