+++ src/main/java/ChatManagement/kafka/application/dto/type/LogType.java
@@ -1,0 +1,5 @@
+package ChatManagement.kafka.application.dto.type;
+
+public enum LogType {
+    INFO, ERROR
+}
