--- src/main/java/ChatManagement/kafka/service/dto/type/LogType.java
@@ -1,5 +1,0 @@
-package ChatManagement.kafka.service.dto.type;
-
-public enum LogType {
-    INFO, ERROR
-}
